Changes to fix problem with rocksdb.rocksdb test on Asan build
Summary:
On Jenkins one of the test cases in the rocksdb.rocksdb
test fails sometimes for the Asan build.  The theory is that a
previous drop table is the cause.  When a table is dropped, the
indexes are put into a list to handle later.  Once all the keys
for the index have moved out of the database (through compaction)
the last information is officially dropped.  Since this operation
does some key updates and occurs in the background it can change
the num_keys_written statistic that we are using for the test
case in question.  In order to avoid it I moved the test case
to the beginning of the file and added an option to force a fresh
restart when running the test.

Differential Revision: https://reviews.facebook.net/D46527

fbshipit-source-id: b298df2be81