Collect statistics in AddUserKey() only for a subset of keys
Summary:
Load testing exposed an issue in `MyRocksTablePropertiesCollector::AddUserKey()` where collecting statistics caused a significant impact during loading data.

Fix is to execute the costly code path only for a certain percentage of cases. The `rocksdb-table-stats-sampling-pct `variable controls the setting. By default we now sample only `10%` of the keys. The allowed values are `{ 0, 1..100 }` with `0` meaning to ignore the sampling checks and use all the data.

Differential Revision: https://reviews.facebook.net/D53469

fbshipit-source-id: 218514c2ea7