mysqld sends error 2006 to socket before closing the connection due to timeout
Summary:
Currently in mysqld when a connection is timed out (defined by wait_timeout) mysqld close the socket directly. Clients won't be able to find out the causes of the lost connection. This diff introduces a new feature to mysqld that when the new added global sys var "send_error_before_closing_timed_out_connection" is truned on server error 2006 with message "Connection closed due to timeout." will be pushed into socket before mysqld closing the connection due to timeout. This error will be delivered and stay in the socket buffer on client side after the connection is closed on server side. Client can read this error and find out the reason of the closed connection is timeout. This feature will help clients improving error handling, to distinguish if it was caused by timeout, or other reasons, e.g. crash, etc.

The purpose of forcing the error code to be 2006 is for compatibility since (client) error code 2006 has been used for this error.

A new global sys var "send_error_before_closing_timed_out_connection" is added in this diff, its default value is false so by default the timeout error message won't be sent, we will turn it on for testing and once all the clients can correctly consume this error message, we will set the default value to true.

To test this feature, a new mysqltest command "dump_timed_out_connection_socket_buffer" is added. Use this command to dump the socket buffer on cilent side after a connection is timeout and verify if the error code / message is valid. Example is as below:

set @global.wait_timeout = 3;
set @global.interactive_timeout = 3;

connect (con1,localhost,root,,);

--echo sleeping 6 sec to make sure the connection get timed out and closed by server...
sleep 6;

--echo dump the client side socket buffer and verify error 2006
dump_timed_out_connection_socket_buffer;

Reference patch:
https://github.com/facebook/mysql-5.6/commit/e04c9e1
https://github.com/facebook/mysql-5.6/commit/12eeda9
https://github.com/facebook/mysql-5.6/commit/27c0b0c

Reviewed By: abal147

Differential Revision: D10040191

fbshipit-source-id: 0e1483f32e7