Renable bulk_load and add a readonly sysvar rocksdb_enable_bulk_load
Summary:
This change includes a fix in RocksDB that will allow statistics to be generated for SST files generated via the SstFileWriter class.  It will also correctly finish and add each table when a new one is started.  In addition I added a readonly system variable (rocksdb_enable_bulk_load_api) that defaults to on and can be used to turn off bulk loading if we need to in the future.

update-submodule: rocksdb

Differential Revision: https://reviews.facebook.net/D63183

fbshipit-source-id: da594ba5cfe