Fix the rocksdb.rocksdb test that was failing on killing a connection
Summary: The MDEV-4059 test where a query waiting on a timeout should be able to be killed before the timeout exceeded was occasionally failing.  It turned out that the kill was generating a condition broadcast, but if it snuck in too early the broadcast occurred before the lock was waiting on the signal and the result of the broadcast was lost.  A simple sleep of one second before killing the query should solve the problem

Reviewed By: lth

Differential Revision: https://reviews.facebook.net/D64743

fbshipit-source-id: 4feb852a500