add explicit fsync to fast secondary index creation
Summary:
Due to the large amount of data written to the os due to fast secondary index
creation, we need to add explicit fsyncs after each call to my_write, as
my_write does not guarantee the data to be flushed to disk.  Otherwise,
potentially writeback and page cache stalls may happen when several indexes are
being created concurrently.

Reviewed By: gunnarku

Differential Revision:
D4330649

fbshipit-source-id: 50d2f09

fbshipit-source-id: 55d1638c902