Fix rocksdb.bulk_load on devservers (#550)
Summary:
The rocksdb.bulk_load test creates and accesses some very large tables and doesn't finish in time on a standard Debug build on a devserver.  Reduce the size of the tables so that it will finish.

I reduced the size of the tables from 10,000,000 rows to 5,000,000 rows.  This should still provide sufficient size to validate what we are trying to test without taking so long.
Closes https://github.com/facebook/mysql-5.6/pull/550

Test Plan: MTR

Differential Revision: D4621037

Pulled By: jkedgar

fbshipit-source-id: 9e604ab2396