Add more information to duplicate key error
Summary: The current error information for duplicate key lists the key value and key name.  In some scenarios this may not be sufficient.  Add a variable that, when enabled, includes the tableame for the key and the query text (or up to 512 bytes of it) into the error message.  The new variable, improved_dup_key_error, will change the output error message to contain this new information.

Reviewed By: hermanlee

Differential Revision: D5167148

fbshipit-source-id: 7ef48fb01f8