Port Control Cross-Table Access to 8.0
Summary:
Port our v5.6 sysvar allow_noncurrent_db_rw, which detects when a query
accesses data not in the current database.  Depending on its setting,
when a query tries to read or write to another database, it will:
ON: Allow it.
LOG: Allow it, but log that it occurred.
LOG_WARN: Allow it, log it occurred, and send an error to client.
OFF: Block it, and send an error to the client.

Reference patch: https://github.com/facebook/mysql-5.6/commit/b5f54b4

Reviewed By: hermanlee

Differential Revision: D6071054

fbshipit-source-id: 6e518de7290