Handle rows_query events correctly when filters are specified
Summary:
When database or table filter was specified along with verbose mode,
mysqlbinlog would print rows_query events of filtered transactions. Fixed that
behavior in this change by buffering the rows_query event until the filtering
decision is taken in the following rows or query event.

Note that the "# at <pos>" comment in the output of mysqlbinlog was always
broken for buffered events (like int_var and user_var events). This diff does
not fix that behavior for rows_query. The result is that the "at" comment for
buffered events are printed consecutively before the actual events.

For 8.0: the binlog_gtid.binlog_rows_query_with_filter requires the
--binlog-trx-meta-data patch to be ported, so it will currently fail.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/b45040ad5a

Reviewed By: abhinav04sharma

Differential Revision: D10219146

fbshipit-source-id: 55710fa5a9b