Class to assign monotonically increasing timestamp
Summary:
Port the HLC clock instance from fb-mysql 5.6 to fb-mysql 8.0. This is a direct
port with very few visual changes.

Adds classes and methods to assign a HLC timestamp to a caller (event or a
transaction). Binlog (MYSQL_BIN_LOG) is the maintainer of the clock (since binlog
is the entity assigning timestamps).
This is a nanosecond precision clock and a 64 bit unsigned integer is used to track the timestamp
The 'logical' part is actually simplified by just incrementing the timestamp on a collision (as in two events/txns occuring within the same nanosecond). This should be  a rare occurence and the wall clock should soon catch up with the HLC clock
Setting a minimum bound on hlc is through a global system variable
The maximum drift allowed on HLC when updating it (thorugh minimum_hlc) is controlled by a sysvar and is defaulted to 300 secs

Reviewed By: abhinav04sharma

Differential Revision: D16553200

fbshipit-source-id: a2a136e1106