Port MTS fixes when hlc is enabled
Summary:
This is a direct port of
https://github.com/facebook/mysql-5.6/commit/d6452f0288c956a841d6a5ba255fd6e8bb37f252

MTS imposes restrictions on binlog event ordering within a  trx. It works by
buffering all the events in a trx until it knows the database to which this trx
should be applied - this means it buffers gtid event, query begin event. With
HLC it should also buffer metadata event. This diff fixes the same and updates a
few debug asserts accordingly.

Reviewed By: abhinav04sharma

Differential Revision: D17415806

fbshipit-source-id: 92da53ebb04