Refactored 'main.all_persisted_variables' MTR test case
Now it does not use actual number of variables in the output, so '.result' file does not need to be re-recorded every time this number changes.

On rebase, squash the binlog_persisted and all_persisted_variables changes with an early patch and to rewrite these tests so that they do not need to be adjusted anymore.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1103

Reviewed By: lth

Pulled By: lth