Call NewJemallocNodumpAllocator only if jemalloc is used
Summary:
The server fails to come up if MyRocks is built without jemalloc, but
the rocksdb_cache_dump feature is enabled. For asan/valgrind builds,
jemalloc should be disabled, but the test might still enable
rocksdb_cache_dump. Allow the asan builds to run with the feature
enabled, but generate a warning to the log instead.

Reviewed By: lth

Differential Revision: D20907021

fbshipit-source-id: 9351d66c217