Coallesce small buffers to send them as one
Summary: The new async client function to write a vector of buffers can cause lots of small buffers to go through SSL encoding which can be expensive.  This diff attempts to combine the first portions (which are most likely to be small) into a single buffer.  The original code already had a small buffer to build the header into.  This expands that buffer and then continues copying data into it until the next buffer won't fit.  This diff also avoids adding 0 length buffers which could occur with the original implementation.  Note that this buffer coalescing is disabled when the overall packet is very large and gets turned off part way through if network compression is enabled.

Reviewed By: abal147

Differential Revision: D21867050

fbshipit-source-id: e9869e040a7