Remove handler::keys_to_use_for_scanning
Summary:
The callback is removed by the following 3 diffs:

https://github.com/facebook/mysql-5.6/commit/3154c713d32258ef0b02b9d44d903178d2d6ed9d
https://github.com/facebook/mysql-5.6/commit/ff63653c6fa803628289bd90e6a2d9117dfdea26
https://github.com/facebook/mysql-5.6/commit/77526ad256ec03e6c0c9a5b326964daf709e9bda

And based on those diffs, this callback was never really necessary as the list of indexes are already calculated from Field::part_of_sortkey and index flag HA_READ_ORDER. For MyRocks that pretty much means every single index, and we return key_map_full in 5.6. So remove the already commented out code.

Reviewed By: Pushapgl

Differential Revision: D22891405

fbshipit-source-id: 1dd1a69485d