Better acl lookup
Summary:
During connection, MySQL consults db_cache to look for (ip, user, db) combinations to find access masks, otherwise it'll fallback to using a linear search (due to possible wildcards in db/host) over `acl_dbs` (which is a in-memory version of mysql.db table). In most cases this is acceptable as mysql.db is typically small. However in certain prod situations when there are too many database / user entries in mysql.db, the sequence search might become a bottleneck.

In 5.6 it is much worse that everything is done under a mutex and we have changed it to a R/W lock in the original patch. In 8.0 there is already a R/W lock protecting the cache, but the linear search is still there, and the cache itself become unbounded hash table while 5.6 it is a bounded LRU cache, so there is a new concern that the cache may grow unbounded.

This adds `enable_acl_fast_lookup` sys var that supports O(k) (user, db) -> [hosts] hash table lookup for all (user, db) pairs that doesn't have wildcards, where k is the number of distinct host patterns for this (user, db) pair, without consulting the cache. In most case the k should be sufficiently small that it is basically a constant.

Most of the original acl_fast_lookup related code have been simplified to use more modern data structures in MySQL 8.0 such as `Prealloced_array<T>` and `malloc_unordered_map`, but the overall design remains largely unchanged.

In our production scenario typically there is only one entry for each (user, db) pair and db is almost always specific (except for test_%, but I doubt production cares about it). So in practice for each (user, db) there is only one host entry so the check becomes O(1). So in most cases we don't even need the db_cache.

Also due to concern of unbounded growth in db_cache, I have added the ability to turn off the db_cache as in most production scenarios this cache should no longer be needed when enable_acl_fast_lookup is on. But I'm keeping the cache on by default for the other uncommon case (such as access denied, missing shard, wildcard db pattern, etc).

It is not 100% clear yet if this is truly required in production since we don't have this exact scenario in 8.0 production yet, but given this is a port it seems good to include this for now in case we need it during 8.0 migration (and drop it later if it turns out it is not required, after running in production for a while).

The new counters in the original patch are not ported as they don't seem particularly helpful in the past.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/cd89a9cec08d1e2184d39cf87b554d29baca6abc

Reviewed By: hermanlee

Differential Revision: D23670998

fbshipit-source-id: e451d5334c7