add update_rli_and_mi to set gtid, master log file and pos
Summary:
This change is:
Before a trx is appended to relay log, the logged gtid sets is updated with the gtid and master log file name/position is updated at the end of the every replicate message on the follower received via Raft.

This ensures that SHOW SLAVE STATUS refledcts up to date info(Read_Master_Log_Pos field), but more importantly keeps the previous gtid set correct on the next file during rotation.

Porting notes:
- use channel_map.get_default_channel_mi() instead of active_mi
- use rli lock instead of global sid lock

Port D19252866 and D21601382

Reviewed By: bhatvinay

Differential Revision: D23981456

fbshipit-source-id: eceb209130f