PS-6790 : Introduce reduced doublewrite buffer mode
Goal:
-----
Introduce an option to write only page_ids (space_id, page_num)
in doublewrite buffer and not the actual pages. This detect-only
mode helps to distinguish from known partial-write corruptions vs
disk corruption issues.

This feature will be useful in the Percona XtraDB cluster because,
on page-corruption, PXC nodes can do SST for the crashed node.

Design:
-------
In reduced doublewrite buffer mode, we store only space_id, page_num,
LSN of pages in doublewrite buffer.

Existing variable --innodb-double-write will be re-used. Currently,
ON, OFF are allowed values. We will introduce a new value REDUCED.

Variable will be dynamic but allowed combinations for dynamic mode are:
REDUCED -> ON
ON      -> REDUCED

We will NOT allow the following combinations
OFF     -> ON
OFF     -> REDUCED
ON      -> OFF
REDUCED -> OFF

A separate file will be used for reduced doublewrite buffer mode.
For batch flushes (flush_list and LRU), we will use the same file.
The file name pattern will be #ib_<page_size>_0.bdblwr

.bdblwr -> batch doublewrite

Single page flushes continue to be written in normal mode. ie. the
full page is written as is.

If innodb-doublewrite=REDUCED is used, and on crash recovery, if the
actual page is corrupted, we crash the server