Move thd_wait_end() calls after mutex is released
Summary:
Entering AC while holding a mutex is prone to deadlocks in case threads holding AC slots are trying to acquire this mutex. This change moves calls to `thd_wait_end()` until after the mutex is released, and `thd->current_mutex` is reset. Also added an assert to make sure that AC is not entered while a mutex is held.

Note: the new assert was hit in debug builds because `DEBUG_SYNC` doesn't reset the current mutex/cond information back if it was null.

Reviewed By: lth

Differential Revision: D26137502

fbshipit-source-id: cb940e427b5